perf: Improve rle_id iteration perf and set sorted flags
